Allow float dimensions in plotting APIs

+++ src/liesel/model/model.py
@@ -2692,8 +2692,8 @@ class Model:
         self,
         show: bool = True,
         save_path: str | None | IO = None,
-        width: int = 14,
-        height: int = 10,
+        width: float = 14,
+        height: float = 10,
         prog: Literal[
             "dot", "circo", "fdp", "neato", "osage", "patchwork", "sfdp", "twopi"
         ] = "dot",
@@ -2744,8 +2744,8 @@ class Model:
         self,
         show: bool = True,
         save_path: str | None | IO = None,
-        width: int = 14,
-        height: int = 10,
+        width: float = 14,
+        height: float = 10,
         prog: Literal[
             "dot", "circo", "fdp", "neato", "osage", "patchwork", "sfdp", "twopi"
         ] = "dot",
@@ -2797,8 +2797,8 @@ class Model:
         self,
         show: bool = True,
         save_path: str | None | IO = None,
-        width: int = 14,
-        height: int = 10,
+        width: float = 14,
+        height: float = 10,
         prog: Literal[
             "dot", "circo", "fdp", "neato", "osage", "patchwork", "sfdp", "twopi"
         ] = "dot",
