+++ src/liesel/model/viz.py
@@ -16,8 +16,8 @@ def plot_nodes(
     model,
     show: bool = True,
     save_path: str | None | IO = None,
-    width: int = 14,
-    height: int = 10,
+    width: float = 14,
+    height: float = 10,
     prog: Literal[
         "dot", "circo", "fdp", "neato", "osage", "patchwork", "sfdp", "twopi"
     ] = "dot",
@@ -75,8 +75,8 @@ def plot_vars(
     model,
     show: bool = True,
     save_path: str | None | IO = None,
-    width: int = 14,
-    height: int = 10,
+    width: float = 14,
+    height: float = 10,
     prog: Literal[
         "dot", "circo", "fdp", "neato", "osage", "patchwork", "sfdp", "twopi"
     ] = "dot",
